Minor fixes and cleanup.
