+++ src/api/steam_api_workshop.cc
@@ -13,6 +13,98 @@ namespace greenworks {
 namespace api {
 namespace {
 
+void InitUgcMatchingTypes(v8::Handle<v8::Object> exports) {
+  v8::Local<v8::Object> ugc_matching_type = Nan::New<v8::Object>();
+  SET_TYPE(ugc_matching_type, "Items", k_EUGCMatchingUGCType_Items);
+  SET_TYPE(ugc_matching_type, "ItemsMtx", k_EUGCMatchingUGCType_Items_Mtx);
+  SET_TYPE(ugc_matching_type, "ItemsReadyToUse",
+           k_EUGCMatchingUGCType_Items_ReadyToUse);
+  SET_TYPE(ugc_matching_type, "Collections", k_EUGCMatchingUGCType_Collections);
+  SET_TYPE(ugc_matching_type, "Artwork", k_EUGCMatchingUGCType_Artwork);
+  SET_TYPE(ugc_matching_type, "Videos", k_EUGCMatchingUGCType_Videos);
+  SET_TYPE(ugc_matching_type, "Screenshots", k_EUGCMatchingUGCType_Screenshots);
+  SET_TYPE(ugc_matching_type, "AllGuides", k_EUGCMatchingUGCType_AllGuides);
+  SET_TYPE(ugc_matching_type, "WebGuides", k_EUGCMatchingUGCType_WebGuides);
+  SET_TYPE(ugc_matching_type, "IntegratedGuides",
+           k_EUGCMatchingUGCType_IntegratedGuides);
+  SET_TYPE(ugc_matching_type, "UsableInGame",
+           k_EUGCMatchingUGCType_UsableInGame);
+  SET_TYPE(ugc_matching_type, "ControllerBindings",
+           k_EUGCMatchingUGCType_ControllerBindings);
+  Nan::Persistent<v8::Object> constructor;
+  constructor.Reset(ugc_matching_type);
+  Nan::Set(exports,
+           Nan::New("UGCMatchingType").ToLocalChecked(),
+           ugc_matching_type);
+}
+
+void InitUgcQueryTypes(v8::Handle<v8::Object> exports) {
+  v8::Local<v8::Object> ugc_query_type = Nan::New<v8::Object>();
+  SET_TYPE(ugc_query_type, "RankedByVote", k_EUGCQuery_RankedByVote);
+  SET_TYPE(ugc_query_type, "RankedByPublicationDate",
+           k_EUGCQuery_RankedByPublicationDate);
+  SET_TYPE(ugc_query_type, "AcceptedForGameRankedByAcceptanceDate",
+           k_EUGCQuery_AcceptedForGameRankedByAcceptanceDate);
+  SET_TYPE(ugc_query_type, "RankedByTrend", k_EUGCQuery_RankedByTrend);
+  SET_TYPE(ugc_query_type, "FavoritedByFriendsRankedByPublicationDate",
+           k_EUGCQuery_FavoritedByFriendsRankedByPublicationDate);
+  SET_TYPE(ugc_query_type, "CreatedByFriendsRankedByPublicationDate",
+           k_EUGCQuery_CreatedByFriendsRankedByPublicationDate);
+  SET_TYPE(ugc_query_type, "RankedByNumTimesReported",
+           k_EUGCQuery_RankedByNumTimesReported);
+  SET_TYPE(ugc_query_type, "CreatedByFollowedUsersRankedByPublicationDate",
+           k_EUGCQuery_CreatedByFollowedUsersRankedByPublicationDate);
+  SET_TYPE(ugc_query_type, "NotYetRated",  k_EUGCQuery_NotYetRated);
+  SET_TYPE(ugc_query_type, "RankedByTotalVotesAsc",
+           k_EUGCQuery_RankedByTotalVotesAsc);
+  SET_TYPE(ugc_query_type, "RankedByVotesUp", k_EUGCQuery_RankedByVotesUp);
+  SET_TYPE(ugc_query_type, "RankedByTextSearch",
+           k_EUGCQuery_RankedByTextSearch);
+  Nan::Persistent<v8::Object> constructor;
+  constructor.Reset(ugc_query_type);
+  Nan::Set(exports,
+           Nan::New("UGCQueryType").ToLocalChecked(),
+           ugc_query_type);
+}
+
+void InitUserUgcList(v8::Handle<v8::Object> exports) {
+  v8::Local<v8::Object> ugc_list = Nan::New<v8::Object>();
+  SET_TYPE(ugc_list, "Published", k_EUserUGCList_Published);
+  SET_TYPE(ugc_list, "VotedOn", k_EUserUGCList_VotedOn);
+  SET_TYPE(ugc_list, "VotedUp", k_EUserUGCList_VotedUp);
+  SET_TYPE(ugc_list, "VotedDown", k_EUserUGCList_VotedDown);
+  SET_TYPE(ugc_list, "WillVoteLater", k_EUserUGCList_WillVoteLater);
+  SET_TYPE(ugc_list, "Favorited", k_EUserUGCList_Favorited);
+  SET_TYPE(ugc_list, "Subscribed", k_EUserUGCList_Subscribed);
+  SET_TYPE(ugc_list, "UsedOrPlayer", k_EUserUGCList_UsedOrPlayed);
+  SET_TYPE(ugc_list, "Followed", k_EUserUGCList_Followed);
+  Nan::Persistent<v8::Object> constructor;
+  constructor.Reset(ugc_list);
+  Nan::Set(exports, Nan::New("UserUGCList").ToLocalChecked(), ugc_list);
+}
+
+void InitUserUgcListSortOrder(v8::Handle<v8::Object> exports) {
+  v8::Local<v8::Object> ugc_list_sort_order = Nan::New<v8::Object>();
+  SET_TYPE(ugc_list_sort_order, "CreationOrderDesc",
+           k_EUserUGCListSortOrder_CreationOrderDesc);
+  SET_TYPE(ugc_list_sort_order, "CreationOrderAsc",
+           k_EUserUGCListSortOrder_CreationOrderAsc);
+  SET_TYPE(ugc_list_sort_order, "TitleAsc", k_EUserUGCListSortOrder_TitleAsc);
+  SET_TYPE(ugc_list_sort_order, "LastUpdatedDesc",
+           k_EUserUGCListSortOrder_LastUpdatedDesc);
+  SET_TYPE(ugc_list_sort_order, "SubscriptionDateDesc",
+           k_EUserUGCListSortOrder_SubscriptionDateDesc);
+  SET_TYPE(ugc_list_sort_order, "VoteScoreDesc",
+           k_EUserUGCListSortOrder_VoteScoreDesc);
+  SET_TYPE(ugc_list_sort_order, "ForModeration",
+           k_EUserUGCListSortOrder_ForModeration);
+  Nan::Persistent<v8::Object> constructor;
+  constructor.Reset(ugc_list_sort_order);
+  Nan::Set(exports,
+           Nan::New("UserUGCListSortOrder").ToLocalChecked(),
+           ugc_list_sort_order);
+}
+
 NAN_METHOD(FileShare) {
   Nan::HandleScope scope;
 
@@ -216,10 +308,10 @@ NAN_METHOD(UGCUnsubscribe) {
 }
 
 void RegisterAPIs(v8::Handle<v8::Object> exports) {
-  utils::InitUgcMatchingTypes(exports);
-  utils::InitUgcQueryTypes(exports);
-  utils::InitUserUgcListSortOrder(exports);
-  utils::InitUserUgcList(exports);
+  InitUgcMatchingTypes(exports);
+  InitUgcQueryTypes(exports);
+  InitUserUgcListSortOrder(exports);
+  InitUserUgcList(exports);
 
   Nan::Set(exports,
            Nan::New("fileShare").ToLocalChecked(),
